+++ src/Core/Serialization/DOM/Normalizers/LicenseNormalizer.php
@@ -55,7 +55,14 @@ class LicenseNormalizer extends _BaseNormalizer
         );
         \assert(null !== $element);
 
-        return $element;
+        return SimpleDOM::setAttributes(
+            $element,
+            [
+                'acknowledgement' => $this->getNormalizerFactory()->getSpec()->supportsLicenseAcknowledgement()
+                    ? $license->getAcknowledgement()?->value
+                    : null,
+            ]
+        );
     }
 
     /**
@@ -76,7 +83,13 @@ class LicenseNormalizer extends _BaseNormalizer
         $document = $factory->getDocument();
 
         return SimpleDOM::appendChildren(
-            $document->createElement('license'),
+            SimpleDOM::setAttributes(
+                $document->createElement('license'),
+                [
+                    'acknowledgement' => $this->getNormalizerFactory()->getSpec()->supportsLicenseAcknowledgement()
+                        ? $license->getAcknowledgement()?->value
+                        : null,
+                ]),
             [
                 SimpleDOM::makeSafeTextElement($document, 'id', $id),
                 SimpleDOM::makeSafeTextElement($document, 'name', $name),