+++ src/Core/Serialization/JSON/Normalizers/LicenseNormalizer.php
@@ -47,7 +47,15 @@ class LicenseNormalizer extends _BaseNormalizer
         // TODO: IMPLEMENTED IF NEEDED: may throw, if not supported by the spec
         // $this->getNormalizerFactory()->getSpec()->supportsLicenseExpression()
 
-        return ['expression' => $license->getExpression()];
+        return array_filter(
+            [
+                'expression' => $license->getExpression(),
+                'acknowledgement' => $this->getNormalizerFactory()->getSpec()->supportsLicenseAcknowledgement()
+                    ? $license->getAcknowledgement()?->value
+                    : null,
+            ],
+            Predicate::isNotNull(...)
+        );
     }
 
     /**
@@ -68,6 +76,9 @@ class LicenseNormalizer extends _BaseNormalizer
                 'id' => $id,
                 'name' => $name,
                 'url' => JsonHelper::encodeIriReferenceBE($license->getUrl()),
+                'acknowledgement' => $this->getNormalizerFactory()->getSpec()->supportsLicenseAcknowledgement()
+                    ? $license->getAcknowledgement()?->value
+                    : null,
             ],
             Predicate::isNotNull(...)
         )];