+++ tests/Core/Models/License/NamedLicenseTest.php
@@ -23,6 +23,7 @@ declare(strict_types=1);
 
 namespace CycloneDX\Tests\Core\Models\License;
 
+use CycloneDX\Core\Enums\LicenseAcknowledgement;
 use CycloneDX\Core\Models\License\_DisjunctiveLicenseBase;
 use CycloneDX\Core\Models\License\NamedLicense;
 use PHPUnit\Framework\Attributes\CoversClass;
@@ -49,6 +50,7 @@ class NamedLicenseTest extends TestCase
         $license = new NamedLicense('');
         self::assertSame('', $license->getName());
         self::assertNull($license->getUrl());
+        self::assertNull($license->getAcknowledgement());
 
         return $license;
     }
@@ -89,5 +91,16 @@ class NamedLicenseTest extends TestCase
     {
         $license->setUrl('');
         self::assertNull($license->getUrl());
+    }
+
+    #[DependsUsingShallowClone('testConstruct')]
+    public function testSetAndGetAcknowledgement(NamedLicense $license): void
+    {
+        $acknowledgement = LicenseAcknowledgement::Declared;
+
+        $got = $license->setAcknowledgement($acknowledgement);
+
+        self::assertSame($license, $got);
+        self::assertSame($acknowledgement, $license->getAcknowledgement());
     }
 }