+++ HISTORY.md
@@ -6,6 +6,17 @@ All notable changes to this project will be documented in this file.
 
 <!-- add unreleased items here -->
 
+* Added
+  * Namespace `\CycloneDX\Core\Enums`
+    * New enum: `LicenseAcknowledgement` ([#428] via [#429])
+  * Namespace `\CycloneDX\Core\Models\License`
+    * New methods `LicenseExpression::{get,set}Acknowledgement()` ([#428] via [#429])
+    * New methods `NamedLicense::{get,set}Acknowledgement()` ([#428] via [#429])
+    * New methods `SpdxLicense::{get,set}Acknowledgement()` ([#428] via [#429])
+
+[#428]: https://github.com/CycloneDX/cyclonedx-php-library/issues/428
+[#429]: https://github.com/CycloneDX/cyclonedx-php-library/pull/429
+
 ## 3.2.0 - 2024-04-10
 
 Added _basic_ support for [_CycloneDX_ Specification-1.6](https://github.com/CycloneDX/specification/releases/tag/1.6).